test: move the new validator tests into the tree CI actually runs
codecov/patch failed at 13.33% of diff hit, which turned out to be the useful
signal: the tests were added to tests/litellm_utils_tests/, and NO workflow runs
that directory. Every test-unit job enumerates paths under tests/test_litellm/,
so the new tests never executed in CI and the new code showed as uncovered.

They passed locally, which is exactly what made it easy to miss -- a regression
test that CI never runs protects nothing.

Moved to tests/test_litellm/test_utils.py, which the `misc` job picks up via its
tests/test_litellm/test_*.py entry. tests/litellm_utils_tests/test_utils.py is
restored byte-identical to upstream, so this PR no longer touches it at all.

  tests/test_litellm/test_utils.py -> 255 passed (5 new, 250 pre-existing)
  ruff format --config ruff.toml -> no hunks touch the moved block
